Move getFilesOfInterest into the IDE, and OfInterest out into DAML
